+++ Source/DynamicScene/GeoJsonDataSource.js
@@ -263,26 +263,26 @@ define([
         //default line
         var defaultLine = new DynamicObject('GeoJsonDataSource.defaultLine');
         var polyline = new DynamicPolyline();
-        polyline.color = new ConstantProperty(Color.YELLOW);
+        var material = new ColorMaterialProperty();
+        material.color = new ConstantProperty(Color.YELLOW);
+        polyline.material = material;
         polyline.width = new ConstantProperty(2);
-        polyline.outlineColor = new ConstantProperty(Color.BLACK);
-        polyline.outlineWidth = new ConstantProperty(1);
         defaultLine.polyline = polyline;
 
         //default polygon
         var defaultPolygon = new DynamicObject('GeoJsonDataSource.defaultPolygon');
 
         polyline = new DynamicPolyline();
-        polyline.color = new ConstantProperty(Color.YELLOW);
+        material = new ColorMaterialProperty();
+        material.color = new ConstantProperty(Color.YELLOW);
+        polyline.material = material;
         polyline.width = new ConstantProperty(1);
-        polyline.outlineColor = new ConstantProperty(Color.BLACK);
-        polyline.outlineWidth = new ConstantProperty(0);
         defaultPolygon.polyline = polyline;
 
         var polygon = new DynamicPolygon();
         defaultPolygon.polygon = polygon;
 
-        var material = new ColorMaterialProperty();
+        material = new ColorMaterialProperty();
         material.color = new ConstantProperty(new Color(1.0, 1.0, 0.0, 0.1));
         polygon.material = material;