+++ Source/Renderer/Context.js
@@ -2545,7 +2545,7 @@ define([
      *
      * @memberof Context
      *
-     * @see Context#pick
+     * @see Scene#pick
      */
     Context.prototype.createPickFramebuffer = function() {
         return new PickFramebuffer(this);