+++ Specs/DynamicScene/CzmlDataSourceSpec.js
@@ -1227,10 +1227,10 @@ defineSuite([
         var dynamicObject = dataSource.getDynamicObjectCollection().getObjects()[0];
 
         expect(dynamicObject.polyline).toBeDefined();
-        expect(dynamicObject.polyline.color.getValue(Iso8601.MINIMUM_VALUE)).toEqual(new Color(0.1, 0.1, 0.1, 0.1));
+        expect(dynamicObject.polyline.material.color.getValue(Iso8601.MINIMUM_VALUE)).toEqual(new Color(0.1, 0.1, 0.1, 0.1));
         expect(dynamicObject.polyline.width.getValue(Iso8601.MINIMUM_VALUE)).toEqual(polylinePacket.polyline.width);
-        expect(dynamicObject.polyline.outlineColor.getValue(Iso8601.MINIMUM_VALUE)).toEqual(new Color(0.2, 0.2, 0.2, 0.2));
-        expect(dynamicObject.polyline.outlineWidth.getValue(Iso8601.MINIMUM_VALUE)).toEqual(polylinePacket.polyline.outlineWidth);
+        expect(dynamicObject.polyline.material.outlineColor.getValue(Iso8601.MINIMUM_VALUE)).toEqual(new Color(0.2, 0.2, 0.2, 0.2));
+        expect(dynamicObject.polyline.material.outlineWidth.getValue(Iso8601.MINIMUM_VALUE)).toEqual(polylinePacket.polyline.outlineWidth);
         expect(dynamicObject.polyline.show.getValue(Iso8601.MINIMUM_VALUE)).toEqual(true);
     });
 
@@ -1258,16 +1258,14 @@ defineSuite([
         var dynamicObject = dataSource.getDynamicObjectCollection().getObjects()[0];
 
         expect(dynamicObject.polyline).toBeDefined();
-        expect(dynamicObject.polyline.color.getValue(validTime)).toEqual(new Color(0.1, 0.1, 0.1, 0.1));
+        expect(dynamicObject.polyline.material.getValue(validTime).color).toEqual(new Color(0.1, 0.1, 0.1, 0.1));
         expect(dynamicObject.polyline.width.getValue(validTime)).toEqual(polylinePacket.polyline.width);
-        expect(dynamicObject.polyline.outlineColor.getValue(validTime)).toEqual(new Color(0.2, 0.2, 0.2, 0.2));
-        expect(dynamicObject.polyline.outlineWidth.getValue(validTime)).toEqual(polylinePacket.polyline.outlineWidth);
+        expect(dynamicObject.polyline.material.getValue(validTime).outlineColor).toEqual(new Color(0.2, 0.2, 0.2, 0.2));
+        expect(dynamicObject.polyline.material.getValue(validTime).outlineWidth).toEqual(polylinePacket.polyline.outlineWidth);
         expect(dynamicObject.polyline.show.getValue(validTime)).toEqual(true);
 
-        expect(dynamicObject.polyline.color.getValue(invalidTime)).toBeUndefined();
+        expect(dynamicObject.polyline.material.getValue(invalidTime)).toBeUndefined();
         expect(dynamicObject.polyline.width.getValue(invalidTime)).toBeUndefined();
-        expect(dynamicObject.polyline.outlineColor.getValue(invalidTime)).toBeUndefined();
-        expect(dynamicObject.polyline.outlineWidth.getValue(invalidTime)).toBeUndefined();
         expect(dynamicObject.polyline.show.getValue(invalidTime)).toBeUndefined();
     });
 
@@ -1495,5 +1493,34 @@ defineSuite([
 
         var child2 = objects.getById('child2');
         expect(child2.parent).toBe(parent);
+    });
+
+    it('Processes JulianDate packets.', function() {
+        var date = JulianDate.fromIso8601('2000-01-01');
+
+        var object = {};
+        CzmlDataSource.processPacketData(JulianDate, object, 'simpleDate', date.toIso8601());
+
+        expect(object.simpleDate).toBeDefined();
+        expect(object.simpleDate.getValue()).toEqual(date);
+
+        CzmlDataSource.processPacketData(JulianDate, object, 'objDate', {
+            date : date.toIso8601()
+        });
+
+        expect(object.objDate).toBeDefined();
+        expect(object.objDate.getValue()).toEqual(date);
+    });
+
+    it('Processes array packets.', function() {
+        var arrayPacket = {
+            array : [1, 2, 3, 4, 5]
+        };
+
+        var object = {};
+        CzmlDataSource.processPacketData(Array, object, 'arrayData', arrayPacket);
+
+        expect(object.arrayData).toBeDefined();
+        expect(object.arrayData.getValue()).toEqual(arrayPacket.array);
     });
 });